Initialized repository for project in augmenting sentiment analysis with social features.
The project leverages the Yelp Academic Dataset and investigates whether adding features from that datasets social graph to standard bag-of-words models for classification can improve performance in sentiment classification.  At present the project implements a baseline logistic regression classifier with three classes: negative, neutral, and positive.  The baseline achieves an average f1 score of 0.77.  With the Yelp dataset downloaded and unzipped, supplying the paths to the review and user json files as arguments to runbaseline.py will read the appropriate data and run the baseline classifier.