Implemented additional baseline classifier based on graph influence.  Altered the way in which data is split into training and testing sets.
cleanyelp.py now contains a method for calculating the median date in the datset.  This is used to split the data chronologically into sets of approximately equal size.  The influence baseline examines influencer reviews for each reveiw in the test set and assigns a label to the test review based on the majority label of the influencers.