Implemented additional methods for exploring data and altered the cleaning of the review dataset.
Added a method to find the median date in a set of reviews.  This may be useful for splitting the data chronologically and could be adapted to split on kth order date.  Implemented find_influencers, which, given a review, finds all reviews of the same business by friends of the user.  Data cleansing now uses this method to remove from the dataset any reviews that do not have any such influencer reviews.