Updated baselineclassifer.py to run several baseline tests.
A naive model of link strength between reviews of the same business by a pair of friends, which assigns a label based on the majority classification among friends who had previously reviewed the business is implemented.  This is coupled with an added randomized class assignment and naive Bayes classifier where there is no freind review to base classification on from the training set.  Naive influence improves classification over randomization, but hurts performance when coupled with naive Bayes.