Altered bag-of-words classifier to return class probabilities for each review.
Rather than a list of predicted labels, the method bag_of_words_probabilities returns a dictionary that maps tuples of the form (r_id, c) to the probability that review denoted by r_id belongs to class c, as determined by the classifier.  This will be useful in building a graph of reviews and classes and defining the edge weights according to the values of these probabilities.