Implemented Linear-chain CRF.
reviewcrf.py implements the training of a Linear-chain CRF model and the tagging of test reviews using pycrfsuite, a python wrapper for the popular CRFSuite fast CRF implementation from Naoaki Okazaki.  Training takes pairs of review of the same business by friends and labels the later review as belonging to either the same sentiment class as the earlier review ("1") or the opposite class ("0").  Current attributes are limited to date and number of mutual friends between users.