Using pygraph, implemented graph formulation to solve optimal classification problem using maximum flow/min-cut.
reviewgraph.py builds a directed graph whose nodes consist of reviews in the test set.  The source node is a node for the "negative" class and the sink node is a node for the "positive" class.  Edges connecting review nodes to class nodes have weights as determined by the bag of words classifier in baselineclassifier.py.  Review nodes for reviews of the same business by friends are connected by edges with weights determined by the Linear-chain CRF implementation.  Classification is performed using the Edmonds-Karp max-flow/min-cut algorithm.