Experimented with different feature functions in CRF, to no positive effect.
Working with features involving cosine similarity of documents to one another or to each class does not seem to help performance.  The CRF outputs that all pairs of reviews should receive the same classification.  I believe that this is largely due to lack of training examples from reviews sequences.  I will try working with CRF training on a sequence of all reviews of the same business.  Further development of feature functions for the CRF will also be worthwhile.